+++ ergo-core/src/main/scala/org/ergoplatform/network/message/NipopowProofData.scala
@@ -1,0 +1,8 @@
+package org.ergoplatform.network.message
+
+import org.ergoplatform.settings.Algos
+import scorex.util.ModifierId
+
+case class NipopowProofData(m: Int, k: Int, headerId: Option[ModifierId]) {
+  def headerIdBytesOpt: Option[Array[Byte]] = headerId.map(Algos.decode).flatMap(_.toOption)
+}
